Integrate Codegen.php into GeneratesColumnsEnum.php
