--- src/Schema/AST/Codegen.php
@@ -1,63 +1,0 @@
-<?php
-
-namespace Nuwave\Lighthouse\Schema\AST;
-
-use GraphQL\Language\AST\EnumTypeDefinitionNode;
-use GraphQL\Language\AST\FieldDefinitionNode;
-use GraphQL\Language\AST\InputValueDefinitionNode;
-use Illuminate\Support\Str;
-
-class Codegen
-{
-    /**
-     * Create the name for the Enum that holds the allowed columns.
-     *
-     * @example FieldNameArgNameColumn
-     *
-     * @param  \GraphQL\Language\AST\InputValueDefinitionNode  $argDefinition
-     * @param  \GraphQL\Language\AST\FieldDefinitionNode  $parentField
-     * @return string
-     */
-    public static function allowedColumnsEnumName(InputValueDefinitionNode &$argDefinition, FieldDefinitionNode &$parentField): string
-    {
-        return Str::studly($parentField->name->value)
-            .Str::studly($argDefinition->name->value)
-            .'Column';
-    }
-
-    /**
-     * Create the Enum that holds the allowed columns.
-     *
-     * @param  \GraphQL\Language\AST\InputValueDefinitionNode  $argDefinition
-     * @param  \GraphQL\Language\AST\FieldDefinitionNode  $parentField
-     * @param  string[]  $allowedColumns
-     * @param  string  $allowedColumnsEnumName
-     * @return \GraphQL\Language\AST\EnumTypeDefinitionNode
-     */
-    public static function createAllowedColumnsEnum(
-        InputValueDefinitionNode &$argDefinition,
-        FieldDefinitionNode &$parentField,
-        array $allowedColumns,
-        string $allowedColumnsEnumName
-    ): EnumTypeDefinitionNode {
-        $enumValues = array_map(
-            function (string $columnName): string {
-                return
-                    strtoupper(
-                        Str::snake($columnName)
-                    )
-                    .' @enum(value: "'.$columnName.'")';
-            },
-            $allowedColumns
-        );
-
-        $enumDefinition = "\"Allowed column names for the `{$argDefinition->name->value}` argument on the query `{$parentField->name->value}`.\"\n"
-            ."enum $allowedColumnsEnumName {\n";
-        foreach ($enumValues as $enumValue) {
-            $enumDefinition .= "$enumValue\n";
-        }
-        $enumDefinition .= '}';
-
-        return PartialParser::enumTypeDefinition($enumDefinition);
-    }
-}
