+++ src/Support/Traits/GeneratesColumnsEnum.php
@@ -2,20 +2,27 @@
 
 namespace Nuwave\Lighthouse\Support\Traits;
 
+use GraphQL\Language\AST\EnumTypeDefinitionNode;
 use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\InputValueDefinitionNode;
+use Illuminate\Support\Str;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
-use Nuwave\Lighthouse\Schema\AST\Codegen;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
+use Nuwave\Lighthouse\Schema\AST\PartialParser;
 
+/**
+ * Directives may want to constrain database columns to an enum.
+ *
+ * @mixin \Nuwave\Lighthouse\Schema\Directives\BaseDirective
+ */
 trait GeneratesColumnsEnum
 {
     /**
-     * Check whether the directive has a list of allowed columns.
+     * Check whether the directive constrains allowed columns.
      *
      * @return bool
      */
-    public function hasAllowedColumns(): bool
+    protected function hasAllowedColumns(): bool
     {
         $hasColumns = ! is_null($this->directiveArgValue('columns'));
         $hasColumnsEnum = ! is_null($this->directiveArgValue('columnsEnum'));
@@ -35,9 +42,9 @@ trait GeneratesColumnsEnum
      * @param \Nuwave\Lighthouse\Schema\AST\DocumentAST $documentAST
      * @param \GraphQL\Language\AST\InputValueDefinitionNode $argDefinition
      * @param \GraphQL\Language\AST\FieldDefinitionNode $parentField
-     * @return string
+     * @return string The name of the used enum.
      */
-    public function generateColumnsEnum(
+    protected function generateColumnsEnum(
         DocumentAST &$documentAST,
         InputValueDefinitionNode &$argDefinition,
         FieldDefinitionNode &$parentField
@@ -48,11 +55,11 @@ trait GeneratesColumnsEnum
             return $columnsEnum;
         }
 
-        $allowedColumnsEnumName = Codegen::allowedColumnsEnumName($argDefinition, $parentField);
+        $allowedColumnsEnumName = static::allowedColumnsEnumName($argDefinition, $parentField);
 
         $documentAST
             ->setTypeDefinition(
-                Codegen::createAllowedColumnsEnum(
+                static::createAllowedColumnsEnum(
                     $argDefinition,
                     $parentField,
                     $this->directiveArgValue('columns'),
@@ -61,5 +68,59 @@ trait GeneratesColumnsEnum
             );
 
         return $allowedColumnsEnumName;
+    }
+
+    /**
+     * Create the name for the Enum that holds the allowed columns.
+     *
+     * @example FieldNameArgNameColumn
+     *
+     * @param  \GraphQL\Language\AST\InputValueDefinitionNode  $argDefinition
+     * @param  \GraphQL\Language\AST\FieldDefinitionNode  $parentField
+     * @return string
+     */
+    protected function allowedColumnsEnumName(
+        InputValueDefinitionNode &$argDefinition,
+        FieldDefinitionNode &$parentField
+    ): string {
+        return Str::studly($parentField->name->value)
+            .Str::studly($argDefinition->name->value)
+            .'Column';
+    }
+
+    /**
+     * Create the Enum that holds the allowed columns.
+     *
+     * @param  \GraphQL\Language\AST\InputValueDefinitionNode  $argDefinition
+     * @param  \GraphQL\Language\AST\FieldDefinitionNode  $parentField
+     * @param  string[]  $allowedColumns
+     * @param  string  $allowedColumnsEnumName
+     * @return \GraphQL\Language\AST\EnumTypeDefinitionNode
+     */
+    protected function createAllowedColumnsEnum(
+        InputValueDefinitionNode &$argDefinition,
+        FieldDefinitionNode &$parentField,
+        array $allowedColumns,
+        string $allowedColumnsEnumName
+    ): EnumTypeDefinitionNode {
+        $enumValues = array_map(
+            function (string $columnName): string {
+                return
+                    strtoupper(
+                        Str::snake($columnName)
+                    )
+                    .' @enum(value: "'.$columnName.'")';
+            },
+            $allowedColumns
+        );
+
+        $enumDefinition = "\"Allowed column names for the `{$argDefinition->name->value}` argument on the query `{$parentField->name->value}`.\"\n"
+            ."enum $allowedColumnsEnumName {\n";
+        foreach ($enumValues as $enumValue) {
+            $enumDefinition .= "$enumValue\n";
+        }
+        $enumDefinition .= '}';
+
+        return PartialParser::enumTypeDefinition($enumDefinition);
     }
 }
